Move methods to separate folder
