+++ families/seth/rpc/src/calls/network.rs
@@ -17,7 +17,7 @@
 
 use jsonrpc_core::{Params, Value, Error};
 
-use super::client::{ValidatorClient};
+use client::{ValidatorClient};
 
 use sawtooth_sdk::messaging::stream::MessageSender;
